fix(ci): 修复 CI 全链失败——删孤儿 command-classes 门禁、补 Docker libclang/tauri core 构建、清 98 处 clippy 与 9 处前端 lint
- 删 command-classes 门禁(清单已在 1bd6cd82 有意删除,门禁守空):
  Makefile target、CI step、scripts/check-command-classes.sh、route_contract Test A
- Dockerfile: rquickjs-sys 的 bindgen 需要 libclang,builder 阶段补装
- ci.yml tauri-rust job: tauri.conf.json 把 core/dist/index.js 列进 resources,
  先 build core 再 cargo check
- backend clippy 98 errors: 契约层模块级 allow + 机械修复
- 前端 biome: 9 个 format/organizeImports error 修复;
  settings/index.ts 是 core 逐字镜像,豁免 organizeImports;
  镜像 getActiveAgentPrompt 补齐 settings 契约

+++ crates/backend/src/services/tunnel/data_plane.rs
@@ -583,7 +583,7 @@ async fn bridge_websocket(
                         .map_err(|e| format!("转发 pong 失败：{e}"))?,
                     AxumMessage::Close(frame) => {
                         let close = frame.map(|f| CloseFrame {
-                            code: CloseCode::from(u16::from(f.code)),
+                            code: CloseCode::from(f.code),
                             reason: f.reason.as_str().into(),
                         });
                         let _ = up_tx.send(WsMessage::Close(close)).await;
